+++ examples/ReferenceTest/test_reference.py
@@ -13,7 +13,7 @@ from sdc11073.xml_types import pm_types, msg_types, pm_qnames as pm
 from sdc11073.definitions_sdc import SDC_v1_Definitions
 from sdc11073.xml_types.dpws_types import ThisDeviceType, ThisModelType
 from sdc11073.location import SdcLocation
-from sdc11073.mdib import DeviceMdibContainer, ClientMdibContainer
+from sdc11073.mdib import ProviderMdibContainer, ClientMdibContainer
 from sdc11073.sdcclient import SdcClient
 from sdc11073.provider.sdcdeviceimpl import SdcDevice
 from sdc11073.wsdiscovery import WSDiscoveryWhitelist, Scopes
@@ -97,7 +97,7 @@ class DeviceActivity(threading.Thread):
 
 
 def createReferenceDevice(wsdiscovery_instance, location, mdibPath):
-    my_mdib = DeviceMdibContainer.from_mdib_file(mdibPath)
+    my_mdib = ProviderMdibContainer.from_mdib_file(mdibPath)
     my_uuid = uuid.UUID(My_Dev_UUID_str)
     dpwsModel = ThisModelType(manufacturer='sdc11073',
                               manufacturer_url='www.sdc11073.com',
