+++ tests/test_device_waveform.py
@@ -5,7 +5,7 @@ import sdc11073
 from sdc11073.xml_types import pm_types
 from sdc11073.xml_types.dpws_types import ThisModelType, ThisDeviceType
 from sdc11073.mdib import descriptorcontainers as dc
-from sdc11073.mdib.devicewaveform import Annotator
+from sdc11073.mdib.providerwaveform import Annotator
 from sdc11073.provider import waveforms
 from sdc11073.pysoap.soapclientpool import SoapClientPool
 from tests import mockstuff
@@ -17,7 +17,7 @@ HANDLES = ("0x34F05506", "0x34F05501", "0x34F05500")
 class TestDeviceWaveform(unittest.TestCase):
 
     def setUp(self):
-        self.mdib = sdc11073.mdib.DeviceMdibContainer()
+        self.mdib = sdc11073.mdib.ProviderMdibContainer()
         self._soap_client_pool = SoapClientPool(soap_client_factory=None, log_prefix="")
 
         # this structure is not realistic, but sufficient for what we need here.
