+++ tests/test_mdib.py
@@ -3,7 +3,7 @@ import unittest
 
 from sdc11073.xml_types import pm_qnames as pm
 from sdc11073.exceptions import ApiUsageError
-from sdc11073.mdib import DeviceMdibContainer
+from sdc11073.mdib import ProviderMdibContainer
 from sdc11073.xml_types.pm_types import Coding
 
 mdib_folder = os.path.dirname(__file__)
@@ -16,7 +16,7 @@ class TestMdib(unittest.TestCase):
 
     def test_select_descriptors(self):
 
-        device_mdib_container = DeviceMdibContainer.from_mdib_file(mdib_70041_path)
+        device_mdib_container = ProviderMdibContainer.from_mdib_file(mdib_70041_path)
         # from looking at the mdib file I know how many elements the tested paths shall return
         for path, expectedCount in [(('70041',), 1),
                                     (('70041', '69650'), 1),  # VMDs
@@ -32,12 +32,12 @@ class TestMdib(unittest.TestCase):
     def test_mdib_tns(self):
         # verify that a mdib with participant model as default namespace can be handled.
         # if creation does not raise any exception, all should be fine.
-        device_mdib_container = DeviceMdibContainer.from_mdib_file(
+        device_mdib_container = ProviderMdibContainer.from_mdib_file(
             os.path.join(os.path.dirname(__file__), 'mdib_tns.xml'))
         self.assertTrue(device_mdib_container is not None)
 
     def test_get_metric_descriptor_by_code(self):
-        device_mdib_container = DeviceMdibContainer.from_mdib_file(mdib_tns_path)
+        device_mdib_container = ProviderMdibContainer.from_mdib_file(mdib_tns_path)
         metric_container = device_mdib_container.get_metric_descriptor_by_code(vmd_code=Coding("130536"),
                                                                                channel_code=Coding("130637"),
                                                                                metric_code=Coding("196174"))
@@ -55,7 +55,7 @@ class TestMdib(unittest.TestCase):
 class TestMdibTransaction(unittest.TestCase):
 
     def setUp(self):
-        self.mdib = DeviceMdibContainer.from_mdib_file(mdib_tns_path)
+        self.mdib = ProviderMdibContainer.from_mdib_file(mdib_tns_path)
 
     def test_create_descriptor(self):
         with self.mdib.transaction_manager() as mgr:
